backend/requirements.txt
ai-edge-litert==1.1.0
pillow==10.4.0
numpy==1.26.4